pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zs-config"
version = "1.1.0"
description = "Interactive TUI for Zscaler OneAPI automation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.33.0",
    "truststore>=0.9.0",
    "zscaler-sdk-python>=1.9.21",
    "rich>=13.7.0",
    "questionary>=2.0.0",
    "pyfiglet>=1.0.2",
    "sqlalchemy>=2.0.0",
    "cryptography>=46.0.6",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
]

[project.scripts]
zs-config = "cli.z_config:main"

[tool.setuptools.packages.find]
include = ["cli*", "db*", "lib*", "services*", "api*"]
